Structure blog roll as more traditional blog
- /blog loads all blog posts in order
  - summary field is used for post preview
  - we can consider adding pagination once there are enough posts
- /blog/:id continues to link to the posts
- blog posts no longer have a dropdown menu, table of contents or whitepaper navigation UI
