+++ src/components/DropdownNav.tsx
@@ -1,0 +1,27 @@
+import React from 'react'
+import { NavDropdown } from 'react-bootstrap'
+
+import { PostMap } from 'posts/post'
+
+export const DropdownNav: React.FunctionComponent<{
+  title: string
+  baseNavPath: string
+  posts: PostMap
+  enableAll: boolean
+}> = ({ title, baseNavPath, posts, enableAll }) => {
+  return (
+    <NavDropdown title={title} id="basic-nav-dropdown">
+      <NavDropdown.Item href={baseNavPath}>Table of Contents</NavDropdown.Item>
+      {enableAll ? (
+        <NavDropdown.Item href={`${baseNavPath}/all`}>
+          All In One Page
+        </NavDropdown.Item>
+      ) : null}
+      {Object.keys(posts).map((pageId, pageIndex) => (
+        <NavDropdown.Item href={`${baseNavPath}/${pageId}`} key={pageIndex}>
+          {pageIndex + 1}. {posts[pageId].shortTitle}
+        </NavDropdown.Item>
+      ))}
+    </NavDropdown>
+  )
+}
