+++ src/components/calculator/FirstTimeUserIntroduction.tsx
