+++ src/posts/blog/index.ts
@@ -1,7 +1,7 @@
-import post1 from 'posts/blog/1-ux-redesign/ux-redesign'
-import { PostMap } from 'posts/post'
+import { post as uxredesign } from 'posts/blog/1-ux-redesign/ux-redesign'
+import { BlogPostMap } from 'posts/post'
 
 // Note: the keys in this map are the public URL for the post. Changing them will break links.
-export const pages: PostMap = {
-  uxredesign: post1,
+export const pages: BlogPostMap = {
+  uxredesign,
 }
