+++ src/posts/paper/12-calculator-and-conclusion.ts
@@ -22,10 +22,8 @@ Thanks for reading!
 
 If you've found this paper or the calculator valuable, consider supporting the project with a small donation. We are all unpaid volunteers and we don't currently have any funding to
 work on this project. Thank you in advance!
-<!-- Donation -->
-
 `
 
-const post = { title, shortTitle, content }
+const post = { title, shortTitle, content, donation: true }
 
 export default post
